Validate role admin

+++ routes/upload.js
@@ -114,10 +114,9 @@ function saveFile(model, type, id, fileName, path, res){
 
     var lastPath = `./uploads/${type}/${model.img}`;
 
-    if(model.img.length > 0 && fs.existsSync(lastPath)){
+    if(model.img && model.img.length > 0 && fs.existsSync(lastPath)){
         fs.unlink(lastPath, err => {
             if (err){
-                console.log('remove 1')
                 return removeFile(path, err, res);
             }
             updateType(model, res);
